move plugin endpoints under connector and processor service
