+++ pkg/web/api/processor_v1.go
@@ -15,11 +15,16 @@
 //go:generate mockgen -destination=mock/processor.go -package=mock -mock_names=ProcessorOrchestrator=ProcessorOrchestrator . ProcessorOrchestrator
 //go:generate mockgen -destination=mock/processor_service_in.go -package=mock -mock_names=ProcessorService_InspectProcessorInServer=ProcessorService_InspectProcessorInServer github.com/conduitio/conduit/proto/api/v1 ProcessorService_InspectProcessorInServer
 //go:generate mockgen -destination=mock/processor_service_out.go -package=mock -mock_names=ProcessorService_InspectProcessorOutServer=ProcessorService_InspectProcessorOutServer github.com/conduitio/conduit/proto/api/v1 ProcessorService_InspectProcessorOutServer
+//go:generate mockgen -destination=mock/processor_plugin.go -package=mock -mock_names=ProcessorPluginOrchestrator=ProcessorPluginOrchestrator . ProcessorPluginOrchestrator
 
 package api
 
 import (
 	"context"
+	"regexp"
+	"slices"
+
+	processorSdk "github.com/conduitio/conduit-processor-sdk"
 
 	"github.com/conduitio/conduit/pkg/foundation/cerrors"
 	"github.com/conduitio/conduit/pkg/inspector"
@@ -48,14 +53,25 @@ type ProcessorOrchestrator interface {
 	InspectOut(ctx context.Context, id string) (*inspector.Session, error)
 }
 
+type ProcessorPluginOrchestrator interface {
+	// List will return all processor plugins' specs.
+	List(ctx context.Context) (map[string]processorSdk.Specification, error)
+}
+
 type ProcessorAPIv1 struct {
 	apiv1.UnimplementedProcessorServiceServer
-	ps ProcessorOrchestrator
+	processorOrchestrator       ProcessorOrchestrator
+	processorPluginOrchestrator ProcessorPluginOrchestrator
 }
 
 // NewProcessorAPIv1 returns a new processor API server.
-func NewProcessorAPIv1(ps ProcessorOrchestrator) *ProcessorAPIv1 {
-	return &ProcessorAPIv1{ps: ps}
+func NewProcessorAPIv1(
+	po ProcessorOrchestrator,
+	ppo ProcessorPluginOrchestrator) *ProcessorAPIv1 {
+	return &ProcessorAPIv1{
+		processorOrchestrator:       po,
+		processorPluginOrchestrator: ppo,
+	}
 }
 
 // Register registers the service in the server.
@@ -67,11 +83,11 @@ func (p *ProcessorAPIv1) ListProcessors(
 	ctx context.Context,
 	req *apiv1.ListProcessorsRequest,
 ) (*apiv1.ListProcessorsResponse, error) {
-	list := p.ps.List(ctx)
+	list := p.processorOrchestrator.List(ctx)
 	var plist []*apiv1.Processor
 
 	for _, v := range list {
-		if len(req.ParentIds) == 0 || p.containsString(req.ParentIds, v.Parent.ID) {
+		if len(req.ParentIds) == 0 || slices.Contains(req.ParentIds, v.Parent.ID) {
 			plist = append(plist, toproto.Processor(v))
 		}
 	}
@@ -87,7 +103,7 @@ func (p *ProcessorAPIv1) InspectProcessorIn(
 		return status.ProcessorError(cerrors.ErrEmptyID)
 	}
 
-	session, err := p.ps.InspectIn(server.Context(), req.Id)
+	session, err := p.processorOrchestrator.InspectIn(server.Context(), req.Id)
 	if err != nil {
 		return status.ProcessorError(cerrors.Errorf("failed to inspect processor: %w", err))
 	}
@@ -117,7 +133,7 @@ func (p *ProcessorAPIv1) InspectProcessorOut(
 		return status.ProcessorError(cerrors.ErrEmptyID)
 	}
 
-	session, err := p.ps.InspectOut(server.Context(), req.Id)
+	session, err := p.processorOrchestrator.InspectOut(server.Context(), req.Id)
 	if err != nil {
 		return status.ProcessorError(cerrors.Errorf("failed to inspect processor: %w", err))
 	}
@@ -149,7 +165,7 @@ func (p *ProcessorAPIv1) GetProcessor(
 	}
 
 	// fetch the processor from the ProcessorOrchestrator
-	pr, err := p.ps.Get(ctx, req.Id)
+	pr, err := p.processorOrchestrator.Get(ctx, req.Id)
 	if err != nil {
 		return nil, status.ProcessorError(cerrors.Errorf("failed to get processor by ID: %w", err))
 	}
@@ -175,7 +191,7 @@ func (p *ProcessorAPIv1) CreateProcessor(
 		plugin = req.Type
 	}
 
-	created, err := p.ps.Create(
+	created, err := p.processorOrchestrator.Create(
 		ctx,
 		plugin,
 		fromproto.ProcessorParent(req.Parent),
@@ -200,7 +216,7 @@ func (p *ProcessorAPIv1) UpdateProcessor(
 		return nil, cerrors.ErrEmptyID
 	}
 
-	updated, err := p.ps.Update(ctx, req.Id, fromproto.ProcessorConfig(req.Config))
+	updated, err := p.processorOrchestrator.Update(ctx, req.Id, fromproto.ProcessorConfig(req.Config))
 
 	if err != nil {
 		return nil, status.ProcessorError(cerrors.Errorf("failed to update processor: %w", err))
@@ -212,7 +228,7 @@ func (p *ProcessorAPIv1) UpdateProcessor(
 }
 
 func (p *ProcessorAPIv1) DeleteProcessor(ctx context.Context, req *apiv1.DeleteProcessorRequest) (*apiv1.DeleteProcessorResponse, error) {
-	err := p.ps.Delete(ctx, req.Id)
+	err := p.processorOrchestrator.Delete(ctx, req.Id)
 
 	if err != nil {
 		return nil, status.ProcessorError(cerrors.Errorf("failed to delete processor: %w", err))
@@ -221,11 +237,31 @@ func (p *ProcessorAPIv1) DeleteProcessor(ctx context.Context, req *apiv1.DeleteP
 	return &apiv1.DeleteProcessorResponse{}, nil
 }
 
-func (p *ProcessorAPIv1) containsString(a []string, s string) bool {
-	for _, v := range a {
-		if v == s {
-			return true
+func (p *ProcessorAPIv1) ListProcessorPlugins(
+	ctx context.Context,
+	req *apiv1.ListProcessorPluginsRequest,
+) (*apiv1.ListProcessorPluginsResponse, error) {
+	var nameFilter *regexp.Regexp
+	if req.GetName() != "" {
+		var err error
+		nameFilter, err = regexp.Compile("^" + req.GetName() + "$")
+		if err != nil {
+			return nil, status.PluginError(cerrors.New("invalid name regex"))
 		}
 	}
-	return false
+
+	mp, err := p.processorPluginOrchestrator.List(ctx)
+	if err != nil {
+		return nil, status.PluginError(err)
+	}
+	var plist []*apiv1.ProcessorPluginSpecifications
+
+	for name, v := range mp {
+		if nameFilter != nil && !nameFilter.MatchString(name) {
+			continue // don't add to result list, filter didn't match
+		}
+		plist = append(plist, toproto.ProcessorPluginSpecifications(name, v))
+	}
+
+	return &apiv1.ListProcessorPluginsResponse{Plugins: plist}, nil
 }
